LayoutTests/http/tests/security/xssAuditor/javascript-link-HTML-entities-null-char-expected.txt

CONSOLE MESSAGE: line 1: SyntaxError: Parse error
